refactor project to feature based architecture
