+++ src/assets/mod.rs
@@ -1,9 +1,11 @@
 pub mod loader;
 pub mod material;
 pub mod mesh;
+pub mod pools;
 pub mod texture;
 
 pub use loader::load_model;
 pub use material::{Material, MaterialHandle};
 pub use mesh::{Mesh, MeshHandle, Primitive};
+pub use pools::{MaterialPool, MeshPool, TexturePool};
 pub use texture::Texture;
