+++ src/camera/mod.rs
@@ -1,4 +1,8 @@
-use bevy_ecs::resource::Resource;
+use bevy_ecs::prelude::*;
+use winit::keyboard::KeyCode;
+
+use crate::input::Input;
+use crate::time::Time;
 
 #[derive(Resource)]
 pub struct Camera {
@@ -36,5 +40,30 @@ impl Camera {
     pub fn view_matrix(&self) -> glam::Mat4 {
         let target = self.position + self.forward();
         glam::Mat4::look_at_rh(self.position, target, glam::Vec3::Y)
+    }
+}
+
+pub fn fly_camera(input: Res<Input>, time: Res<Time>, mut camera: ResMut<Camera>) {
+    let forward_dir = camera.forward();
+    let right_dir = camera.right();
+    let speed = camera.speed * time.delta_time;
+
+    if input.pressed.contains(&KeyCode::KeyW) {
+        camera.position += forward_dir * speed;
+    }
+    if input.pressed.contains(&KeyCode::KeyS) {
+        camera.position -= forward_dir * speed;
+    }
+    if input.pressed.contains(&KeyCode::KeyD) {
+        camera.position += right_dir * speed;
+    }
+    if input.pressed.contains(&KeyCode::KeyA) {
+        camera.position -= right_dir * speed;
+    }
+    if input.pressed.contains(&KeyCode::Space) {
+        camera.position.y += speed;
+    }
+    if input.pressed.contains(&KeyCode::ShiftLeft) {
+        camera.position.y -= speed;
     }
 }
