+++ src/input/mod.rs
@@ -1,4 +1,4 @@
-use bevy_ecs::resource::Resource;
+use bevy_ecs::prelude::*;
 use std::collections::HashSet;
 use winit::keyboard::KeyCode;
 
@@ -6,4 +6,8 @@ use winit::keyboard::KeyCode;
 pub struct Input {
     pub pressed: HashSet<KeyCode>,
     pub just_pressed: HashSet<KeyCode>,
+}
+
+pub fn clear_just_pressed(mut input: ResMut<Input>) {
+    input.just_pressed.clear();
 }
