+++ src/render/mod.rs
@@ -1,9 +1,11 @@
 pub mod depth;
 pub mod instance;
+pub mod prepare;
 pub mod uniforms;
 pub mod vertex;
 
 pub use depth::create_depth_texture;
 pub use instance::{INSTANCE_BUFFER_INITIAL_CAPACITY, InstanceData};
+pub use prepare::prepare_renderables;
 pub use uniforms::{CameraUniform, LightUniform};
 pub use vertex::Vertex;
