+++ src/render/uniforms.rs
@@ -14,5 +14,5 @@ pub struct LightUniform {
 #[derive(Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
 pub struct CameraUniform {
     pub view_proj: [[f32; 4]; 4],
-    pub camera_pos: [f32; 4], // w unused
+    pub position: [f32; 4], // w unused
 }
