--- src/resources/mod.rs
@@ -1,13 +1,0 @@
-pub mod camera;
-pub mod helmet;
-pub mod input;
-pub mod pools;
-pub mod runtime_helmets;
-pub mod time;
-
-pub use camera::Camera;
-pub use helmet::HelmetHandles;
-pub use input::Input;
-pub use pools::{MaterialPool, MeshPool, TexturePool};
-pub use runtime_helmets::RuntimeHelmets;
-pub use time::Time;
