--- src/systems/mod.rs
@@ -1,10 +1,0 @@
-pub mod camera;
-pub mod input;
-pub mod prepare;
-pub mod spawn;
-pub mod time;
-
-pub use camera::fly_camera;
-pub use input::clear_just_pressed;
-pub use spawn::{despawn_helmet, spawn_helmet};
-pub use time::tick_time;
