--- src/systems/time.rs
@@ -1,7 +1,0 @@
-use bevy_ecs::system::ResMut;
-
-use crate::resources::Time;
-
-pub fn tick_time(mut time: ResMut<Time>) {
-    time.tick();
-}
